+++ Stitch/Graph/StitchAI/Mapping/ActionsToSyntax/StitchToPatchData.swift
@@ -75,80 +75,3 @@ extension String {
             .replacingOccurrences(of: "-", with: "_")
     }
 }
-
-
-// TODO: move
-
-extension Array where Element == NodeConnectionType {
-    func createSwiftUICodeArgs(varIdNameMap: [UUID: String],
-                               isLayer: Bool) throws -> [String] {
-        try self.map { inputData in
-            try inputData.createSwiftUICodeArg(varIdNameMap: varIdNameMap,
-                                               isLayer: isLayer)
-        }
-    }
-    
-    func createSwiftUICodeArgs(patchNodeEntityMap: [UUID: PatchNodeEntity]) throws -> [String] {
-        try self.map { inputData in
-            try inputData.createSwiftUICodeArg(patchNodeEntityMap: patchNodeEntityMap)
-        }
-    }
-}
-
-extension NodeConnectionType {
-    /// Called for patches, used for creating SwiftUI code from a Stitch graph.
-    func createSwiftUICodeArg(patchNodeEntityMap: [UUID: PatchNodeEntity]) throws -> String {
-        switch self {
-        case .values(let values):
-            return try values.createSwiftUICodeArg()
-            
-        case .upstreamConnection(let upstream):
-            guard let upstreamPatchNode = patchNodeEntityMap.get(upstream.nodeId),
-                  let portIndex = upstream.portId else {
-                throw SwiftUISyntaxError.upstreamVarNameNotFound(upstream)
-            }
-            
-            let upstreamVarName = upstreamPatchNode.patch.rawValue.createUniqueVarName(nodeId: upstream.nodeId)
-            
-            // Port indices used just for patches
-            return "\(upstreamVarName)[\(portIndex)]"
-        }
-    }
-    
-    func createSwiftUICodeArg(varIdNameMap: [UUID: String],
-                              isLayer: Bool) throws -> String {
-        switch self {
-        case .values(let values):
-            return try values.createSwiftUICodeArg()
-            
-        case .upstreamConnection(let upstream):
-            // Variable name should already exist given topological order, otherwise its a cycle case which we should ignore
-            guard let upstreamVarName = varIdNameMap.get(upstream.nodeId),
-                  let portIndex = upstream.portId else {
-                throw SwiftUISyntaxError.upstreamVarNameNotFound(upstream)
-            }
-            
-            // Port indices used just for patches
-            if isLayer {
-                return "\(upstreamVarName)[\(portIndex)]"
-            }
-            
-            return upstreamVarName
-        }
-    }
-}
-
-extension Array where Element == PortValue {
-    func createSwiftUICodeArg() throws -> String {
-        guard let firstValue = self.first else {
-            fatalError()
-        }
-        
-        let valueDesc = PrintablePortValueDescription(firstValue)
-        let string = try valueDesc.jsonWithoutQuotedKeys()
-        
-        // gets rid of brackets
-        let trimmedStr = string.dropFirst().dropLast()
-        return "[PortValueDescription(\(trimmedStr))]"
-    }
-}
